examples: move runc-hook to container-hook
The two following examples were part duplicates:
- runc-hook
- container-hook

The container-hook example was less complete.

This patch does:
- removing container-hook
- port/rename runc-hook to container-hook
