+++ .github/workflows/inspektor-gadget.yml
@@ -961,7 +961,7 @@ jobs:
     strategy:
       fail-fast: false
       matrix:
-        example: [runc-hook, kube-container-collection]
+        example: [container-hook, kube-container-collection]
     steps:
     - uses: actions/checkout@11bd71901bbe5b1630ceea73d27597364c9af683 # v4.2.2
     - name: Set up Docker Buildx
