+++ examples/container-hook/.gitignore
@@ -1,0 +1,2 @@
+container-hook
+container-hook-static
