+++ examples/container-hook/Dockerfile
@@ -6,8 +6,8 @@ COPY examples/go.mod examples/go.sum /gadget/examples/
 RUN cd /gadget && go mod download
 
 COPY ./ /gadget
-RUN cd /gadget/examples/runc-hook && make runc-hook-static
+RUN cd /gadget/examples/container-hook && make container-hook-static
 
 FROM busybox@sha256:9ae97d36d26566ff84e8893c64a6dc4fe8ca6d1144bf5b87b2b85a32def253c7
-COPY --from=builder /gadget/examples/runc-hook/runc-hook-static /bin/runc-hook
-ENTRYPOINT ["/bin/runc-hook"]
+COPY --from=builder /gadget/examples/container-hook/container-hook-static /bin/container-hook
+ENTRYPOINT ["/bin/container-hook"]
