+++ examples/container-hook/Dockerfile.dockerignore
