--- examples/runc-hook/.gitignore
@@ -1,2 +1,0 @@
-runc-hook
-runc-hook-static
